Moving some functions to game_utils.  Adding error return for some of the cmd_moves functions.
